warn

+++ framework/src/interface/operation/operation_impl.h
@@ -210,7 +210,7 @@ constexpr uint32_t PAD_TAIL_INDEX = 5;
 constexpr uint32_t PAD_STRIDE_H = 0;
 constexpr uint32_t PAD_STRIDE_W = 1;
 constexpr uint32_t PAD_STRIDE_D = 2;
-constexpr uint32_t MAX_LOOP = 2000;
+constexpr size_t MAX_LOOP = 2000;
 
 const std::string OP_ATTR_PREFIX = "op_attr_";
 const std::string CONV_PADDINGS_ATTR = OP_ATTR_PREFIX + "paddings";
